backend/requirements.txt
fastapi
uvicorn
pymupdf
google-generativeai
python-multipart
pydantic
python-dotenv
